@ejona86 Broken out from headers canonicalization fix