@louiscryan 2015-11-13
Some benchmarks

Before 
```
Benchmark                                   (channelCount)  (maxConcurrentStreams)   Mode  Cnt       Score      Error  Units
UnaryCallQpsBenchmark.unary:callsPerSecond               1                      10  thrpt   20   12701.774 ±  672.164  ops/s
UnaryCallQpsBenchmark.unary:callsPerSecond               1                     100  thrpt   20   50460.543 ± 1991.761  ops/s
UnaryCallQpsBenchmark.unary:callsPerSecond               1                    1000  thrpt   20   43994.030 ± 4627.185  ops/s
UnaryCallQpsBenchmark.unary:callsPerSecond               2                      10  thrpt   20   25394.865 ± 1172.532  ops/s
UnaryCallQpsBenchmark.unary:callsPerSecond               2                     100  thrpt   20   83723.793 ± 9044.291  ops/s
UnaryCallQpsBenchmark.unary:callsPerSecond               2                    1000  thrpt   20   73084.762 ± 9515.067  ops/s
UnaryCallQpsBenchmark.unary:callsPerSecond               4                      10  thrpt   20   50828.970 ± 4391.662  ops/s
UnaryCallQpsBenchmark.unary:callsPerSecond               4                     100  thrpt   20  113618.368 ± 1898.532  ops/s
UnaryCallQpsBenchmark.unary:callsPerSecond               4                    1000  thrpt   20   97190.744 ± 2162.832  ops/s
UnaryCallQpsBenchmark.unary:callsPerSecond               8                      10  thrpt   20   81407.111 ± 1622.743  ops/s
UnaryCallQpsBenchmark.unary:callsPerSecond               8                     100  thrpt   20   62781.673 ± 1714.861  ops/s
UnaryCallQpsBenchmark.unary:callsPerSecond               8                    1000  thrpt   20   60119.207 ± 3934.070  ops/s
```
After
```
Benchmark                                   (channelCount)  (maxConcurrentStreams)   Mode  Cnt       Score      Error  Units
UnaryCallQpsBenchmark.unary:callsPerSecond               1                      10  thrpt   20   14115.534 ±  625.781  ops/s
UnaryCallQpsBenchmark.unary:callsPerSecond               1                     100  thrpt   20   46233.564 ± 6020.062  ops/s
UnaryCallQpsBenchmark.unary:callsPerSecond               1                    1000  thrpt   20   51611.494 ± 1194.361  ops/s
UnaryCallQpsBenchmark.unary:callsPerSecond               2                      10  thrpt   20   47194.316 ± 2706.383  ops/s
UnaryCallQpsBenchmark.unary:callsPerSecond               2                     100  thrpt   20   85691.272 ± 6258.795  ops/s
UnaryCallQpsBenchmark.unary:callsPerSecond               2                    1000  thrpt   20   72939.409 ± 8803.213  ops/s
UnaryCallQpsBenchmark.unary:callsPerSecond               4                      10  thrpt   20   45257.520 ± 2507.887  ops/s
UnaryCallQpsBenchmark.unary:callsPerSecond               4                     100  thrpt   20  129795.210 ± 4035.769  ops/s
UnaryCallQpsBenchmark.unary:callsPerSecond               4                    1000  thrpt   20  106075.847 ± 4250.647  ops/s
UnaryCallQpsBenchmark.unary:callsPerSecond               8                      10  thrpt   20  120928.208 ± 2365.616  ops/s
UnaryCallQpsBenchmark.unary:callsPerSecond               8                     100  thrpt   20  155078.165 ± 4040.154  ops/s
UnaryCallQpsBenchmark.unary:callsPerSecond               8                    1000  thrpt   20  105319.533 ± 3551.960  ops/s
```